refactor(resource): move resourceOf method to ResourceOf utils class
