+++ infrastructure/src/main/java/com/kaua/ecommerce/users/infrastructure/api/controllers/AccountController.java
@@ -8,19 +8,15 @@ import com.kaua.ecommerce.users.application.account.retrieve.get.GetAccountByIdC
 import com.kaua.ecommerce.users.application.account.retrieve.get.GetAccountByIdUseCase;
 import com.kaua.ecommerce.users.application.account.update.avatar.UpdateAvatarCommand;
 import com.kaua.ecommerce.users.application.account.update.avatar.UpdateAvatarUseCase;
-import com.kaua.ecommerce.users.domain.utils.Resource;
 import com.kaua.ecommerce.users.infrastructure.accounts.models.CreateAccountApiInput;
 import com.kaua.ecommerce.users.infrastructure.accounts.models.GetAccountPresenter;
 import com.kaua.ecommerce.users.infrastructure.accounts.presenters.AccountApiPresenter;
 import com.kaua.ecommerce.users.infrastructure.api.AccountAPI;
-import com.kaua.ecommerce.users.infrastructure.exceptions.ImageSizeNotValidException;
-import com.kaua.ecommerce.users.infrastructure.exceptions.ImageTypeNotValidException;
+import com.kaua.ecommerce.users.infrastructure.utils.ResourceOf;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
 
 @RestController
 public class AccountController implements AccountAPI {
@@ -68,50 +64,12 @@ public class AccountController implements AccountAPI {
     public ResponseEntity<?> updateAccountAvatar(String id, MultipartFile avatarFile) {
         return ResponseEntity.status(HttpStatus.OK).body(
                 this.updateAvatarUseCase.execute(UpdateAvatarCommand.with(
-                        id, resourceOf(avatarFile))));
+                        id, ResourceOf.with(avatarFile))));
     }
 
     @Override
     public ResponseEntity<Void> deleteAccount(String id) {
         this.deleteAccountUseCase.execute(DeleteAccountCommand.with(id));
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    private Resource resourceOf(final MultipartFile part) {
-        if (part == null) {
-            return null;
-        }
-
-        isValidImage(part);
-
-        try {
-            return Resource.with(
-                    part.getInputStream(),
-                    part.getContentType(),
-                    part.getOriginalFilename()
-            );
-        } catch (final Throwable t) {
-            throw new RuntimeException(t);
-        }
-    }
-
-    private void isValidImage(final MultipartFile part) {
-        final var imageTypes = List.of(
-                "image/jpeg",
-                "image/jpg",
-                "image/png"
-        );
-
-        final var validImageType = imageTypes.contains(part.getContentType());
-        final var IMAGE_SIZE = 600;
-        final var validImageSize = part.getSize() > IMAGE_SIZE * 1024;
-
-        if (!validImageType) {
-            throw new ImageTypeNotValidException();
-        }
-
-        if (validImageSize) {
-            throw new ImageSizeNotValidException();
-        }
     }
 }
